Exercise 2.66 (#139)
* Add lookup function and corresponding tests in SICP Chapter 2, part 3

Added a lookup function in the SICP chapter 2, part 3. This function designed to be applicable when the set of records is structured as a binary tree. Along with this function, corresponding tests have also been introduced to validate its correctness.

* Exercise 2.66

Implemented the lookup procedure as described in exercise 2.66 of SICP Chapter 2, Part 3, for a record set structured as a binary tree. Updated the README.md file to reflect completion of this exercise and devised tests to ensure the correct functioning of the lookup procedure.