Add exercise 2.68 (#142)
* Add exercise 2.68 and its corresponding test

Created new function 'encode-symbol' and 'encode' as per Exercise 2.68 and added corresponding tests in 'ex_2_68_test.clj'. Also, updated the README file to include a reference to exercise 2.68.

* Update required elements from clojure.test in ex_2_68_test

The required elements in the ex_2_68_test.clj file have been modified to specifically include only 'deftest' and 'is' from clojure.test. This was done to make the code cleaner and more efficient.