Add exercise 2.69 (#145)
* Add exercise 2.69

Implemented exercise 2.69 which involves generating a Huffman encoding tree. Updated the make-leaf-set function in the book_2_3.clj file for correct functionality. The README has been updated to reflect the progress of completed exercises and the addition of exercise 2.69.

* Refactor Huffman Tree related tests and logic

Refactored the logic related to Huffman Tree examples and tests. Created constants for symbols, messages, and Huffman Tree in book_2_3.clj and reference these constants in tests to reduce code duplication. This also allows easier changes in future as only the main definition needs to be adjusted.

* Remove unused import and update encode call in test

Removed unused 'huffman-pairs' import from book_2_3_test.clj. Also, updated the call to 'encode' in ex_2_68_test.clj to use the 'huffman-tree' defined in 'b23', providing cleaner and more maintainable code.