Minor fixes code style (#158)
* Correct indentation and rearrange required namespaces

Changes implemented in this commit primarily involve code's visual indentations and the arrangement of namespaces in scripts from chapter 1 and chapter 2 for better readability. No changes were made to the logic of the code, ensuring that the programs function as previously intended.

* Remove unnecessary whitespace in book_2_3.clj

An unnecessary whitespace following the huffman-message-encoded definition in book_2_3.clj was removed. This cleanup is part of ongoing efforts to improve the readability and maintainability of the code.